Adressed GNATS:2422, 2642, 2643, 2633, 2644. Also update planDbWriter to make it more legible
git-svn-id: https://europa-pso.googlecode.com/svn/PLASMA/trunk@2064 23e1bc29-d954-0410-9766-47de34158ba0

+++ CBPlanner/component/PartialPlanWriter.cc
@@ -1088,7 +1088,7 @@ namespace Prototype {
           return NINFINITY;
         else {
           std::stringstream stream;
-          stream << (int) dom.getUpperBound();
+          stream <<  dom.getUpperBound();
           return std::string(stream.str());
         }
       }
@@ -1110,7 +1110,7 @@ namespace Prototype {
         else if((int) dom.getLowerBound() == MINUS_INFINITY)
           return NINFINITY;
         std::stringstream stream;
-        stream << (int) dom.getLowerBound();
+        stream << dom.getLowerBound();
         return std::string(stream.str());
       }
       else if(LabelStr::isString((int)dom.getLowerBound())) {
